feat: AWS connection verification, provision button, log viewer improvements
Worker:
- Add CONNECTION_TEST job type (tests AssumeRole into user's account)
- Write errors to job_logs (stderr) so they appear in the log viewer
- Fix db_min/max_capacity type from *int to *float64 (Supabase stores as numeric)
- Add ngrok-skip-browser-warning header for dev testing
- Add fetchAwsResources for caching VPCs/subnets/hosted zones on connection

Trellis:
- AWS connection flow now verifies via worker (queues CONNECTION_TEST job, polls for result)
- Launch Stack button with pre-filled CloudFormation Quick Create URL
- Provision button on vine detail sheet (queues DEPLOY job, opens log viewer)
- Log viewer: loading/waiting states, job status badge, success/failure footer with error message
- Log viewer: simulate button with realistic 35-step provisioning demo
- getJobStatus server action for polling
- verifyAwsIdentity server action
- provisionConfiguration server action

Infrastructure:
- ECR repository in Terraform (replacing GHCR)
- CONNECTION_TEST job type migration

+++ apps/grape/types/configuration.go
@@ -19,8 +19,8 @@ type Configuration struct {
 	ContainerPlatform       string    `json:"container_platform"`
 	CreateVpc               *bool     `json:"create_vpc"`
 	CreatedAt               time.Time `json:"created_at"`
-	DbMaxCapacity           *int      `json:"db_max_capacity"`
-	DbMinCapacity           *int      `json:"db_min_capacity"`
+	DbMaxCapacity           *float64  `json:"db_max_capacity"`
+	DbMinCapacity           *float64  `json:"db_min_capacity"`
 	Description             *string   `json:"description"`
 	DnsDomainName           *string   `json:"dns_domain_name"`
 	DnsHostedZone           *string   `json:"dns_hosted_zone"`
